backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# Hybrid retrieval
rank-bm25>=0.2.2             # BM25 lexical index (replaces TF-IDF)
sentence-transformers>=2.7.0 # Dense embedding model (all-MiniLM-L6-v2)

# Web content extraction (for URL sources)
httpx>=0.27.0
trafilatura>=1.9.0
beautifulsoup4>=4.12.0

# PDF extraction
pypdf>=4.2.0

# Database (admin / postgres ingestion pipeline)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0            # SQLite async driver for local dev

# Alembic migrations
alembic>=1.13.0
